requirements.txt
python-pptx>=0.6.23
Pillow>=10.0.0
defusedxml>=0.7.1
markitdown>=0.1.0
six>=1.16.0
